refactor: rename blank.rb to core_ext.rb
The file now covers both blank?/present? and deep_dup polyfills,
so core_ext.rb better reflects its scope (mirrors active_support/core_ext convention).
