+++ lib/pact/support/core_ext.rb
